Membuat dialog konfirmasi hapus

+++ app/src/main/java/org/d3if3060/assessment1/ui/screen/DetailScreen.kt
@@ -85,6 +85,10 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
         mutableStateOf(false)
     }
 
+    var showDialog by remember {
+        mutableStateOf(false)
+    }
+
     LaunchedEffect(true) {
         if (id == null) return@LaunchedEffect
         val data = viewModel.getLaporan(id) ?: return@LaunchedEffect
@@ -139,6 +143,13 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
                     }
                     if (id != null) {
                         DeleteAction {
+                            showDialog = true
+                        }
+                        DisplayAlertDialog(
+                            openDialog = showDialog,
+                            onDismissRequest = { showDialog = false  }
+                        ) {
+                            showDialog = false
                             viewModel.delete(id)
                             navController.popBackStack()
                         }
